+++ src/features/about/assets/CONTRIBUTORS_LIST.ts
@@ -1,0 +1,3 @@
+import { CONTRIBUTORS_INFO } from "@/features/about";
+
+export const CONTRIBUTORS_LIST = Object.entries(CONTRIBUTORS_INFO);
