+++ src/features/about/assets/TEAM.ts
@@ -1,0 +1,8 @@
+export const TEAM = [
+  "hdp123",
+  "kbnopro",
+  "unknown1508",
+  "minhnhatnoe",
+  "ducanhgh",
+  "natsukagami",
+] as const;
