+++ src/features/about/assets/index.ts
@@ -1,0 +1,3 @@
+export * from "./CONTRIBUTORS_INFO";
+export * from "./CONTRIBUTORS_LIST";
+export * from "./TEAM";
