+++ src/features/about/index.ts
@@ -1,1 +1,2 @@
+export * from "./assets";
 export * from "./types";
